Remove v2 endpoint
