+++ src/config/urls.py
@@ -16,13 +16,8 @@ urlpatterns_v1 = [
     path("chains/", include("chains.urls", namespace="chains")),
 ]
 
-urlpatterns_v2 = [
-    path("chains/", include("chains.urls_v2", namespace="chains")),
-]
-
 urlpatterns = [
     path("api/v1/", include((urlpatterns_v1, "v1"), namespace="v1")),
-    path("api/v2/", include((urlpatterns_v2, "v2"), namespace="v2")),
     path("admin/", admin.site.urls),
     path("check/", lambda request: HttpResponse("Ok"), name="check"),
     re_path(
